dev into master
dev into master

+++ src/views/cart/cart.js
@@ -231,125 +231,135 @@ function addAllElements() {
 function addAllEvents() {
   allSelectCheckbox.addEventListener("change", toggleAll);
   partialDeleteLabel.addEventListener("click", deleteSelectedItems);
-  purchaseButton.addEventListener("click", navigate("/order"));
+  purchaseButton.addEventListener("click", function(e){
+    const currentCount = convertToNumber(productsCountElem.innerText);
+    if( currentCount === 0 ){
+      alert('구매할 상품이 없습니다.');
+    } else {
+      window.location.href = '/order'; // navigate 함수가 안되어 직접 href 변경하는 식으로 처리해주었습니다.
+    }
+  });
 }
 
 // indexedDB의 cart와 order에서 필요한 정보를 가져온 후
 // 요소(컴포넌트)를 만들어 html에 삽입함.
 async function insertProductsfromCart() {
   const products = await getFromDb("cart");
-  const { selectedIds } = await getFromDb("order", "summary");
+  const summary = await getFromDb("order", "summary");
 
-  products.forEach(async (product) => {
-    // 객체 destructuring
-    const { code, quantity, imageUrl, price } = product;
-    //const imageUrl = await getImageUrl(imageKey);
-
-    const isSelected = selectedIds.includes(code);
-
-    cartProductsContainer.insertAdjacentHTML(
-      "beforeend",
-      `
-        <div class="cart-product-item" id="productItem-${code}">
-          <label class="checkbox">
-            <input type="checkbox" id="checkbox-${code}" ${isSelected ? "checked" : ""
-      } />
-          </label>
-          <button class="delete-button" id="delete-${code}">
-            <span class="icon">
-              <i class="fas fa-trash-can"></i>
-            </span>
-          </button>
-          <figure class="image is-96x96">
-            <img
-              id="image-${code}"
-              src="${imageUrl}"
-              alt="product-image"
-            />
-          </figure>
-          <div class="content">
-            <p id="title-${code}">${product.name}</p>
-            <p style="font-weight:lighter;font-size:small;">[상품옵션: ${product.category} ${product.size}]</p>
-            <div class="quantity">
-              <button 
-                class="button is-rounded" 
-                id="minus-${code}" 
-                ${quantity <= 1 ? "disabled" : ""}
-                ${isSelected ? "" : "disabled"}
-              >
-                <span class="icon is-small">
-                  <i class="fas fa-thin fa-minus"></i>
-                </span>
-              </button>
-              <input
-                class="input"
-                id="quantityInput-${code}"
-                type="number"
-                min="1"
-                max="99"
-                value="${quantity}"
-                ${isSelected ? "" : "disabled"}
+  if( summary) {
+    const selectedIds = summary.selectedIds
+    products.forEach(async (product) => {
+      // 객체 destructuring
+      const { code, quantity, imageUrl, price } = product;
+      //const imageUrl = await getImageUrl(imageKey);
+  
+      const isSelected = selectedIds.includes(code);
+  
+      cartProductsContainer.insertAdjacentHTML(
+        "beforeend",
+        `
+          <div class="cart-product-item" id="productItem-${code}">
+            <label class="checkbox">
+              <input type="checkbox" id="checkbox-${code}" ${isSelected ? "checked" : ""
+        } />
+            </label>
+            <button class="delete-button" id="delete-${code}">
+              <span class="icon">
+                <i class="fas fa-trash-can"></i>
+              </span>
+            </button>
+            <figure class="image is-96x96">
+              <img
+                id="image-${code}"
+                src="${imageUrl}"
+                alt="product-image"
               />
-              <button 
-                class="button is-rounded" 
-                id="plus-${code}"
-                ${quantity >= 99 ? "disabled" : ""}
-                ${isSelected ? "" : "disabled"}
-              >
+            </figure>
+            <div class="content">
+              <p id="title-${code}">${product.name}</p>
+              <p style="font-weight:lighter;font-size:small;">[상품옵션: ${product.category} ${product.size}]</p>
+              <div class="quantity">
+                <button 
+                  class="button is-rounded" 
+                  id="minus-${code}" 
+                  ${quantity <= 1 ? "disabled" : ""}
+                  ${isSelected ? "" : "disabled"}
+                >
+                  <span class="icon is-small">
+                    <i class="fas fa-thin fa-minus"></i>
+                  </span>
+                </button>
+                <input
+                  class="input"
+                  id="quantityInput-${code}"
+                  type="number"
+                  min="1"
+                  max="99"
+                  value="${quantity}"
+                  ${isSelected ? "" : "disabled"}
+                />
+                <button 
+                  class="button is-rounded" 
+                  id="plus-${code}"
+                  ${quantity >= 99 ? "disabled" : ""}
+                  ${isSelected ? "" : "disabled"}
+                >
+                  <span class="icon">
+                    <i class="fas fa-lg fa-plus"></i>
+                  </span>
+                </button>
+              </div>
+            </div>
+            <div class="calculation">
+              <p id="unitPrice-${code}">${addCommas(price)}원</p>
+              <p>
                 <span class="icon">
-                  <i class="fas fa-lg fa-plus"></i>
+                  <i class="fas fa-thin fa-xmark"></i>
                 </span>
-              </button>
+              </p>
+              <p id="quantity-${code}">${quantity}</p>
+              <p>
+                <span class="icon">
+                  <i class="fas fa-thin fa-equals"></i>
+                </span>
+              </p>
+              <p id="total-${code}">${addCommas(quantity * removeCommas(price))}원</p>
             </div>
           </div>
-          <div class="calculation">
-            <p id="unitPrice-${code}">${addCommas(price)}원</p>
-            <p>
-              <span class="icon">
-                <i class="fas fa-thin fa-xmark"></i>
-              </span>
-            </p>
-            <p id="quantity-${code}">${quantity}</p>
-            <p>
-              <span class="icon">
-                <i class="fas fa-thin fa-equals"></i>
-              </span>
-            </p>
-            <p id="total-${code}">${addCommas(quantity * removeCommas(price))}원</p>
-          </div>
-        </div>
-      `
-    );
-
-    // 각종 이벤트 추가
-    document
-      .querySelector(`#delete-${code}`)
-      .addEventListener("click", () => deleteItem(code));
-
-    document
-      .querySelector(`#checkbox-${code}`)
-      .addEventListener("change", () => toggleItem(code));
-
-    document
-      .querySelector(`#image-${code}`)
-      .addEventListener("click", navigate(`/product/detail?id=${code}`));
-
-    document
-      .querySelector(`#title-${code}`)
-      .addEventListener("click", navigate(`/product/detail?id=${code}`));
-
-    document
-      .querySelector(`#plus-${code}`)
-      .addEventListener("click", () => increaseItemQuantity(code));
-
-    document
-      .querySelector(`#minus-${code}`)
-      .addEventListener("click", () => decreaseItemQuantity(code));
-
-    document
-      .querySelector(`#quantityInput-${code}`)
-      .addEventListener("change", () => handleQuantityInput(code));
-  });
+        `
+      );
+  
+      // 각종 이벤트 추가
+      document
+        .querySelector(`#delete-${code}`)
+        .addEventListener("click", () => deleteItem(code));
+  
+      document
+        .querySelector(`#checkbox-${code}`)
+        .addEventListener("change", () => toggleItem(code));
+  
+      document
+        .querySelector(`#image-${code}`)
+        .addEventListener("click", navigate(`/product/detail?id=${code}`));
+  
+      document
+        .querySelector(`#title-${code}`)
+        .addEventListener("click", navigate(`/product/detail?id=${code}`));
+  
+      document
+        .querySelector(`#plus-${code}`)
+        .addEventListener("click", () => increaseItemQuantity(code));
+  
+      document
+        .querySelector(`#minus-${code}`)
+        .addEventListener("click", () => decreaseItemQuantity(code));
+  
+      document
+        .querySelector(`#quantityInput-${code}`)
+        .addEventListener("change", () => handleQuantityInput(code));
+    });
+  }
 }
 
 async function toggleItem(id) {
@@ -369,31 +379,34 @@ async function toggleItem(id) {
 async function toggleAll(e) {
   // 전체 체크냐 전체 체크 해제이냐로 true 혹은 false
   const isCheckAll = e.target.checked;
-  const { ids } = await getFromDb("order", "summary");
+  const summary = await getFromDb("order", "summary");
 
-  ids.forEach(async (id) => {
-    const itemCheckbox = document.querySelector(`#checkbox-${id}`);
-    const isItemCurrentlyChecked = itemCheckbox.checked;
-
-    // 일단 아이템(제품) 체크박스에 전체 체크 혹은 언체크 여부를 반영함.
-    itemCheckbox.checked = isCheckAll;
-
-    // 결제정보 업데이트 필요 여부 확인
-    const isAddRequired = isCheckAll && !isItemCurrentlyChecked;
-    const isRemoveRequired = !isCheckAll && isItemCurrentlyChecked;
-
-    // 결제정보 업데이트 및, 체크 상태에서는 수정 가능으로 함
-    if (isAddRequired) {
-      updateOrderSummary(id, "add-checkbox");
-      setQuantityBox(id, "able");
-    }
-
-    // 결제정보 업데이트 및, 언체크 상태에서는 수정 불가능으로 함
-    if (isRemoveRequired) {
-      updateOrderSummary(id, "removeTemp-checkbox");
-      setQuantityBox(id, "disable");
-    }
-  });
+  if (summary){
+    const ids = summary.ids;
+    ids.forEach(async (id) => {
+      const itemCheckbox = document.querySelector(`#checkbox-${id}`);
+      const isItemCurrentlyChecked = itemCheckbox.checked;
+  
+      // 일단 아이템(제품) 체크박스에 전체 체크 혹은 언체크 여부를 반영함.
+      itemCheckbox.checked = isCheckAll;
+  
+      // 결제정보 업데이트 필요 여부 확인
+      const isAddRequired = isCheckAll && !isItemCurrentlyChecked;
+      const isRemoveRequired = !isCheckAll && isItemCurrentlyChecked;
+  
+      // 결제정보 업데이트 및, 체크 상태에서는 수정 가능으로 함
+      if (isAddRequired) {
+        updateOrderSummary(id, "add-checkbox");
+        setQuantityBox(id, "able");
+      }
+  
+      // 결제정보 업데이트 및, 언체크 상태에서는 수정 불가능으로 함
+      if (isRemoveRequired) {
+        updateOrderSummary(id, "removeTemp-checkbox");
+        setQuantityBox(id, "disable");
+      }
+    });
+  }
 }
 
 async function increaseItemQuantity(id) {
@@ -410,6 +423,28 @@ async function increaseItemQuantity(id) {
 
   // 수량 변경박스(-버튼, 입력칸, +버튼) 상태 업데이트
   setQuantityBox(id, "plus");
+
+  //mongoDB의 수량 업데이트
+  const dbQuantity=parseInt(document.querySelector(`#quantityInput-${id}`).value);
+  //console.log(dbQuantity,typeof dbQuantity);
+  const selectItem = await getFromDb("cart",id);
+  const increase_id=selectItem._id;
+  const data={
+    "quantity":dbQuantity
+  }
+  const bodydata = JSON.stringify(data);
+  try {
+    await fetch(`/api/cart?id=${increase_id}`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${sessionStorage.getItem("token")}`,
+      },
+      body: bodydata,
+    });
+  } catch (err) {
+    console.log('db 수량 증가 실패');
+  }
 }
 
 async function decreaseItemQuantity(id) {
@@ -426,7 +461,30 @@ async function decreaseItemQuantity(id) {
 
   // 수량 변경박스(-버튼, 입력칸, +버튼) 상태 업데이트
   setQuantityBox(id, "minus");
+
+  //mongoDB의 수량 업데이트
+  const dbQuantity=parseInt(document.querySelector(`#quantityInput-${id}`).value);
+  //console.log(`수량:${dbQuantity}`);
+  const selectItem = await getFromDb("cart",id);
+  const decrease_id=selectItem._id;
+  const data={
+    "quantity":dbQuantity
+  }
+  const bodydata = JSON.stringify(data);
+  try {
+    await fetch(`/api/cart?id=${decrease_id}`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${sessionStorage.getItem("token")}`,
+      },
+      body: bodydata,
+    });
+  } catch (err) {
+    console.log('db 수량감소 실패');
+  }
 }
+//test
 
 async function handleQuantityInput(id) {
   // 우선 입력값이 범위 1~99 인지 확인
@@ -450,6 +508,25 @@ async function handleQuantityInput(id) {
 
   // 수량 변경박스(-버튼, 입력칸, +버튼) 상태 업데이트
   setQuantityBox(id, "input");
+
+  //mongoDB의 수량 업데이트
+  console.log(quantity,typeof quantity);
+  const selectItem = await getFromDb("cart",id);
+  const change_id=selectItem._id;
+  const data={quantity}
+  const bodydata = JSON.stringify(data);
+  try {
+    await fetch(`/api/cart?id=${change_id}`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${sessionStorage.getItem("token")}`,
+      },
+      body: bodydata,
+    });
+  } catch (err) {
+    console.log('db 수량 변경 실패');
+  }
 }
 
 // -버튼, 숫자입력칸, +버튼 활성화 여부 및 값을 세팅함.
@@ -509,32 +586,56 @@ function setQuantityBox(id, type) {
 }
 
 async function deleteSelectedItems() {
-  const { selectedIds } = await getFromDb("order", "summary");
-
-  selectedIds.forEach((id) => deleteItem(id));
+  const summary = await getFromDb("order", "summary");
+  if ( summary ){
+    const selectedIds = summary.selectedIds;
+    selectedIds.forEach((id) => deleteItem(id));
+  } else {
+    alert('삭제할 상품이 없습니다.');
+  }
 }
 
 // 전체선택 체크박스를, 현재 상황에 맞추어
 // 체크 또는 언체크 상태로 만듦
 async function updateAllSelectCheckbox() {
-  const { ids, selectedIds } = await getFromDb("order", "summary");
+  const summary = await getFromDb("order", "summary");
 
-  const isOrderEmpty = ids.length === 0;
-  const isAllItemSelected = ids.length === selectedIds.length;
+  if ( summary ){
+    const { ids, selectedIds } = summary;
 
-  // 장바구니 아이템(제품) 수가 0이 아니고,
-  // 또 전체 아이템들이 선택된 상태라면 체크함.
-  if (!isOrderEmpty && isAllItemSelected) {
-    allSelectCheckbox.checked = true;
-  } else {
-    allSelectCheckbox.checked = false;
+    const isOrderEmpty = ids.length === 0;
+    const isAllItemSelected = ids.length === selectedIds.length;
+  
+    // 장바구니 아이템(제품) 수가 0이 아니고,
+    // 또 전체 아이템들이 선택된 상태라면 체크함.
+    if (!isOrderEmpty && isAllItemSelected) {
+      allSelectCheckbox.checked = true;
+    } else {
+      allSelectCheckbox.checked = false;
+    }
   }
 }
 
 async function deleteItem(id) {
+  //mongodb에서 데이터를 삭제함
+  const selectItem = await getFromDb("cart",id);
+  const delete_id=selectItem._id;
+  console.log(delete_id);
+  try{
+    await fetch(`/api/cart/delete?id=${delete_id}`, {
+    method: "DELETE",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${sessionStorage.getItem("token")}`,
+    }
+  });
+  }catch(err){
+    alert(`${_id}를 삭제하는데 오류가 발생했습니다.`)
+  }
+
   // indexedDB의 cart 목록에서 id를 key로 가지는 데이터를 삭제함.
   await deleteFromDb("cart", id);
-
+  
   // 결제정보를 업데이트함.
   await updateOrderSummary(id, "removePermanent-deleteButton");
 
@@ -706,17 +807,26 @@ async function updateProductItem(id, type) {
 
 // 페이지 로드 시 실행되며, 결제정보 카드에 값을 삽입함.
 async function insertOrderSummary() {
-  const { productsCount, productsTotal } = await getFromDb("order", "summary");
+  const summary = await getFromDb("order", "summary");
 
-  const hasItems = productsCount !== 0;
+  if (summary ) {
+    const { productsCount, productsTotal } = summary;
 
-  productsCountElem.innerText = `${productsCount}개`;
-  productsTotalElem.innerText = `${parseInt(removeCommas(productsTotal))}원`;
+    const hasItems = productsCount !== 0;
 
-  if (hasItems) {
-    deliveryFeeElem.innerText = `3,000원`;
-    orderTotalElem.innerText = `${addCommas(parseInt(removeCommas(productsTotal)) + 3000)}원`;
+    productsCountElem.innerText = `${productsCount}개`;
+    productsTotalElem.innerText = `${parseInt(removeCommas(productsTotal))}원`;
+
+    if (hasItems) {
+      deliveryFeeElem.innerText = `3,000원`;
+      orderTotalElem.innerText = `${addCommas(parseInt(removeCommas(productsTotal)) + 3000)}원`;
+    } else {
+      deliveryFeeElem.innerText = `0원`;
+      orderTotalElem.innerText = `0원`;
+    }
   } else {
+    productsCountElem.innerText = '0개';
+    productsTotalElem.innerText = '0원';
     deliveryFeeElem.innerText = `0원`;
     orderTotalElem.innerText = `0원`;
   }
